+++ cechmate/__init__.py
@@ -1,7 +1,7 @@
-"""Cechmate init file."""
-
 # from .filtrations import *
 # from .solver import *
 # from .utils import *
 
-from ._version import __version__ as __version__
+from ._version import __version__
+
+__all__ = ["__version__"]