+++ cechmate/filtrations/cech.py
@@ -1,4 +1,5 @@
 import itertools
+import warnings
 import numpy as np
 
 from .base import BaseFiltration
@@ -35,7 +36,10 @@ class Cech(BaseFiltration):
         simplices:
             Cech filtration for the data X
         """
-
+        warnings.warn(
+            "This function is deprecated and will be removed in a future release. Use fit instead.",
+            DeprecationWarning,
+        )
         N = X.shape[0]
         xr = np.arange(N)
         xrl = xr.tolist()
@@ -52,6 +56,41 @@ class Cech(BaseFiltration):
         for k in range(maxdim + 1):
             for idxs in itertools.combinations(xrl, k + 2):
                 C, r2 = miniball(frozenset(idxs), frozenset([]))
+                simplices.append((list(idxs), np.sqrt(r2)))
+
+        self.simplices_ = simplices
+
+        return simplices
+
+    def fit(self, X) -> list[tuple[list[int], int]]:
+        """Compute the Cech filtration of a Euclidean point set for simplices up to order :code:`self.max_dim`.
+
+        Parameters
+        ===========
+
+        X: Nxd array
+            N Euclidean vectors in d dimensions
+
+        Returns
+        ==========
+
+        simplices:
+            Cech filtration for the data X
+        """
+
+        N = X.shape[0]
+        xr = np.arange(N)
+        xrl = xr.tolist()
+        maxdim = self.maxdim or X.shape[1] - 1
+        miniball = miniball_cache(X)
+
+        # start with vertices
+        simplices = [([i], 0) for i in range(N)]
+
+        # then higher order simplices
+        for k in range(maxdim + 1):
+            for idxs in itertools.combinations(xrl, k + 2):
+                _, r2 = miniball(frozenset(idxs), frozenset([]))
                 simplices.append((list(idxs), np.sqrt(r2)))
 
         self.simplices_ = simplices