+++ test/test_cech.py
@@ -1,32 +1,60 @@
 import pytest
 
 import numpy as np
-from cechmate import Cech
+from cechmate.filtrations import Cech
 
 
 @pytest.fixture
-def triangle():
+def equilateral_triangle():
+    """Define an equilateral triangle to see importance of Cech."""
     x = np.array(
         [
-            [0, 0.0],
-            [1, 1.0],
-            [0, 1.0],
+            [0.0, 0.0, 1.0],
+            [0.0, 1.0, 0.0],
+            [1.0, 0.0, 0.0],
         ]
     )
 
     return x
 
 
-def test_triangle(triangle):
-    """Expect 3 vertices, 3 edges, and a triangle"""
-    c = Cech(2).build(triangle)
+def test_triangle(equilateral_triangle):
+    """Expect 3 vertices, 3 edges, and a triangle."""
+    c = Cech(maxdim=2)
+    c_simplices = c.fit(equilateral_triangle)
 
-    assert len(c) == 7
+    assert len(c_simplices) == 7
 
-    vertices = [s for s in c if len(s[0]) == 1]
-    edges = [s for s in c if len(s[0]) == 2]
-    triangles = [s for s in c if len(s[0]) == 3]
+    vertices = [s for s in c_simplices if len(s[0]) == 1]
+    edges = [s for s in c_simplices if len(s[0]) == 2]
+    triangles = [s for s in c_simplices if len(s[0]) == 3]
 
     assert len(vertices) == 3
     assert len(edges) == 3
     assert len(triangles) == 1
+
+    c_diagrams = c.transform(c_simplices)
+    assert len(c_diagrams) == 2
+    assert len(c_diagrams[0]) == 3
+    assert len(c_diagrams[1]) == 1
+
+
+def test_backwards_compatibility(equilateral_triangle):
+    """Test new fit function aligns with old build function."""
+    old = Cech(maxdim=2).build(equilateral_triangle)
+    new = Cech(maxdim=2).fit(equilateral_triangle)
+    assert len(old) == len(new)
+
+    old_vertices = [s for s in old if len(s[0]) == 1]
+    new_vertices = [s for s in new if len(s[0]) == 1]
+    assert old_vertices == new_vertices
+    old_edges = [s for s in old if len(s[0]) == 2]
+    new_edges = [s for s in new if len(s[0]) == 2]
+    assert old_edges == new_edges
+    old_triangles = [s for s in old if len(s[0]) == 3]
+    new_triangles = [s for s in new if len(s[0]) == 3]
+    assert old_triangles == new_triangles
+
+    assert len(old_vertices) == 3
+    assert len(old_edges) == 3
+    assert len(old_triangles) == 1